factory/pyproject.toml
[project]
name = "notify-factory"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=45.0.5",
    "feedparser>=6.0.11",
    "hydra-core>=1.3.2",
    "omegaconf>=2.3.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "rich>=14.0.0",
    "streamlit>=1.46.1",
    "streamlit-code-editor>=0.1.22",
]
